[#3755] Fix addresbook, able to make synchronous calls to eds

+++ sflphone-client-gnome/src/config/addressbook-config.h
@@ -50,12 +50,12 @@ typedef struct _AddressBook_Config
 {
 	// gint64: a signed integer guaranteed to be 64 bits on all platforms
 	// To print or scan values of this type, use G_GINT64_MODIFIER and/or G_GINT64_FORMAT
-    gint64 enable;
-    gint64 max_results;
-    gint64 display_contact_photo;
-    gint64 search_phone_home;
-    gint64 search_phone_business;
-    gint64 search_phone_mobile;
+    gint enable;
+    gint max_results;
+    gint display_contact_photo;
+    gint search_phone_home;
+    gint search_phone_business;
+    gint search_phone_mobile;
 } AddressBook_Config;
 
 /**
